refactor calendar folders
