+++ src/components/Calendar/GooseDay/GooseDay.jsx
@@ -1,5 +1,5 @@
 import { useParams } from 'react-router-dom';
-import WeekCalendar from '../DayCalendarHead/WeekCalendar';
+import WeekCalendar from '../CalendarHeadDay/WeekCalendar';
 
 export const GooseDay = () => {
   const params = useParams();
