+++ src/components/Calendar/PeriodTypeSelect/StyledPeriodTypeSelect.jsx
